Update dependencies
Most versions were updated to their latest with the exception of `react` and `react-dom`.  This was to satisfy some version constraints imposed by other packages.  Even though everything would likely work nicely with React 17, I upped to the latest version of React 16 to play it safe.

This removes one dependency, `@babel/preset-es2015`.  It gives the following message: "We've deprecated any official yearly presets in 6.x in favor or babel-preset-env. For 7.x it would be @babel/preset-env."

This also adds one dependency, `eslint-plugin-react-hooks`.  This is to satisfy a new constraint imposed by `eslint-config-airbnb`.